Broke state_machine.h into several source/header files
